aplicacion metodo post
se crea el metodo post y boton limpiar

+++ Mobile_App_Estiven/Mobile_App_Estiven/Data/Models/ListUser.cs
@@ -16,7 +16,7 @@ namespace Mobile_App_Estiven.Data.Models
 
         public string Marca { get; set; }
         public int Cantidad { get; set; }
-        public float ValorUnitario { get; set; }
+        public int ValorUnitario { get; set; }
         public long IdUser { get; set; }
 
         public UserDto User { get; set; }
